ensure that the bundled NPM dependencies are only loaded by the background script
this improves the load time of all extension pages
