+++ extension/background/DID.mjs
@@ -1,0 +1,35 @@
+import { base58btc, base64url, ed25519, getPublicKey, varint } from "/external/background.mjs";
+
+export * from "/shared/js/DID.mjs";
+
+const ED25519_CODEC_ID = varint(parseInt("0xed", 16));
+
+/**
+ * Generates a DID.
+ * @return {Object}
+ */
+export async function generateDID() {
+	let privateKeyBytes = ed25519.randomPrivateKey();
+	let publicKeyBytes = await getPublicKey(privateKeyBytes);
+
+	let idBytes = new Uint8Array(publicKeyBytes.byteLength + ED25519_CODEC_ID.length);
+	idBytes.set(ED25519_CODEC_ID, 0);
+	idBytes.set(publicKeyBytes, ED25519_CODEC_ID.length);
+
+	let id = base58btc.encode(idBytes);
+	let did = `did:key:${id}`;
+	let keyId = `${did}#${id}`;
+	let publicJWK = {
+		alg: "EdDSA",
+		crv: "Ed25519",
+		kid: keyId,
+		kty: "OKP",
+		use: "sig",
+		x: base64url.baseEncode(publicKeyBytes)
+	};
+	let privateJWK = {
+		...publicJWK,
+		d: base64url.baseEncode(privateKeyBytes),
+	};
+	return { did, publicJWK, privateJWK };
+}
